moved the production code into the framework
